Enable continuous deployment to Azure
This does several things. First, we define the WEBAPP_NAME as a environment variable, which is the name of the website (pgscm).

Next we create a build job. This will go as follows.
1) First it requires the build job to have completed with passing tests.
2) Then it will run on ubuntu-latest for the environment
3) Then it will download the new artifact: app which in turn will be used to deploy
4) Then it will log into Azure using the azure GitHub Action
5) Then it will disable any static collection and run the migrations for the application
6) Finally it will deploy the app to azure.

This will utilize a new secret: AZURE_SERVICE_PRINCIPAL, which will be used to handle all authentication to the azure site."