Enable pushing to Azure only when the changes are merged into Main
Originally, I had it setup to deploy once tests pass, but it would probably be a bad idea to do that on every PR. We should only save this for when they are officially merged into main.

This creates a new file: python-app-pr.yml that will run our test suite on Pull Requests. Then when the changes are merged into main, the python-app.yml jobs will deploy to azure.