Enable Django rest framework in settings without any functionality
This adds the rest_framework as an installed app.

In addition, it defines the REST_FRAMEWORK variable that contains the permissions for using the API. We are going to require any user attempting to use our API is authenticated in some form.

In the future, we will look to maximize the django rest framework. For now this is just enough to get it configured into the project.