Create a json view for post requests
This will verify an API key of the correct size has been given in the HTTP Request Header. If so, then if will verify it is in the database.

Next it will verify if a student info is valid in the database. Finally, it will save the student if all is correct.