This show the process a user takes when genereting a new api key
First, we test if a modal appears that asks the user if they would like confirm the regeneration of the key.

Secondly, we test when a user cancels a regeneration, then they are redirected to the api profile page.

Finally, we test when a user confirms the regeneration, that the key is actually regenerated.