Delete an api bdd feature
This will first verify a modal pops up when a user wishes to delete an particular api key.

Secondly, this will test if a user cancels the delete confirmation, that they are now on they key's profile page.

Finally, we test if a user confirms the deletion, that the key is not present on the api portal.